Refactor typings to share code
